Web page layout complete.
The web page design is now complete. Ground layout is in place and every css is set up for each element.

Every part is in a jspf element to make it more mudular and to try and ease the process of making it SPA.

The Index.html file is how it looks when put together.